Pipelines/pyproject.toml
[project]
name = "pipelines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "bvcscrap>=0.2.1",
    "feedparser>=6.0.11",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.12",
    "neo4j>=6.1.0",
    "pandas>=3.0.2",
    "pyyaml>=6.0.3",
    "scrapy>=2.14.2",
    "scrapy-playwright>=0.0.46",
    "tavily-python>=0.7.23",
]
